Correct bug when using Tab-mode "size"
Because of 3 misplaced parens in the calculation for how many tabs per dimension that integer value could be a float x.5 instead. Later that would be used in a range statement producing a crash. Also further calcucaltions seem to depend on that value being integer as using int(range) avoids the crash but creates a faulty box.